+++ test/example_domain/money_transfer/transfer_money_process_manager.ex
@@ -1,56 +1,59 @@
 defmodule Commanded.ExampleDomain.TransferMoneyProcessManager do
   @moduledoc false
+
   use Commanded.ProcessManagers.ProcessManager,
-    name: "transfer_money_process_manager",
+    name: __MODULE__,
     router: Commanded.ExampleDomain.BankRouter
 
   @derive Jason.Encoder
-  defstruct transfer_uuid: nil,
-            debit_account: nil,
-            credit_account: nil,
-            amount: nil,
-            status: nil
+  defstruct [:transfer_uuid, :debit_account, :credit_account, :amount, :status]
 
   alias Commanded.ExampleDomain.TransferMoneyProcessManager
-  alias Commanded.ExampleDomain.MoneyTransfer.Events.{MoneyTransferRequested}
-  alias Commanded.ExampleDomain.BankAccount.Events.{MoneyDeposited, MoneyWithdrawn}
-  alias Commanded.ExampleDomain.BankAccount.Commands.{DepositMoney, WithdrawMoney}
+  alias Commanded.ExampleDomain.MoneyTransfer.Events.MoneyTransferRequested
+  alias Commanded.ExampleDomain.BankAccount.Events.MoneyDeposited
+  alias Commanded.ExampleDomain.BankAccount.Events.MoneyWithdrawn
+  alias Commanded.ExampleDomain.BankAccount.Commands.DepositMoney
+  alias Commanded.ExampleDomain.BankAccount.Commands.WithdrawMoney
 
   def interested?(%MoneyTransferRequested{transfer_uuid: transfer_uuid}),
     do: {:start, transfer_uuid}
 
-  def interested?(%MoneyWithdrawn{transfer_uuid: transfer_uuid}), do: {:continue, transfer_uuid}
-  def interested?(%MoneyDeposited{transfer_uuid: transfer_uuid}), do: {:continue, transfer_uuid}
+  def interested?(%MoneyWithdrawn{transfer_uuid: transfer_uuid}),
+    do: {:continue, transfer_uuid}
 
-  def handle(%TransferMoneyProcessManager{}, %MoneyTransferRequested{
-        transfer_uuid: transfer_uuid,
-        debit_account: debit_account,
-        amount: amount
-      }) do
+  def interested?(%MoneyDeposited{transfer_uuid: transfer_uuid}),
+    do: {:continue, transfer_uuid}
+
+  def handle(%TransferMoneyProcessManager{}, %MoneyTransferRequested{} = event) do
+    %MoneyTransferRequested{
+      transfer_uuid: transfer_uuid,
+      debit_account: debit_account,
+      amount: amount
+    } = event
+
     %WithdrawMoney{account_number: debit_account, transfer_uuid: transfer_uuid, amount: amount}
   end
 
-  def handle(
-        %TransferMoneyProcessManager{
-          transfer_uuid: transfer_uuid,
-          credit_account: credit_account,
-          amount: amount
-        },
-        %MoneyWithdrawn{}
-      ) do
+  def handle(%TransferMoneyProcessManager{} = pm, %MoneyWithdrawn{}) do
+    %TransferMoneyProcessManager{
+      transfer_uuid: transfer_uuid,
+      credit_account: credit_account,
+      amount: amount
+    } = pm
+
     %DepositMoney{account_number: credit_account, transfer_uuid: transfer_uuid, amount: amount}
   end
 
-  def handle(%TransferMoneyProcessManager{}, %MoneyDeposited{}), do: []
+  ## State mutators
 
-  ## state mutators
+  def apply(%TransferMoneyProcessManager{} = transfer, %MoneyTransferRequested{} = event) do
+    %MoneyTransferRequested{
+      transfer_uuid: transfer_uuid,
+      debit_account: debit_account,
+      credit_account: credit_account,
+      amount: amount
+    } = event
 
-  def apply(%TransferMoneyProcessManager{} = transfer, %MoneyTransferRequested{
-        transfer_uuid: transfer_uuid,
-        debit_account: debit_account,
-        credit_account: credit_account,
-        amount: amount
-      }) do
     %TransferMoneyProcessManager{
       transfer
       | transfer_uuid: transfer_uuid,