+++ test/helpers/event_factory.ex
@@ -2,7 +2,7 @@ defmodule Commanded.Helpers.EventFactory do
   @moduledoc false
   alias Commanded.EventStore.RecordedEvent
 
-  def map_to_recorded_events(events) do
+  def map_to_recorded_events(events, initial_event_number \\ 1) do
     stream_id = UUID.uuid4()
     causation_id = UUID.uuid4()
     correlation_id = UUID.uuid4()
@@ -10,7 +10,7 @@ defmodule Commanded.Helpers.EventFactory do
 
     events
     |> Commanded.Event.Mapper.map_to_event_data(fields)
-    |> Enum.with_index(1)
+    |> Enum.with_index(initial_event_number)
     |> Enum.map(fn {event, index} ->
       %RecordedEvent{
         event_id: UUID.uuid4(),